Refactor harmony builder as hook
