+++ src/types/interval.ts
@@ -1,0 +1,5 @@
+export type Interval = {
+    semitones: number;
+    ratio: number;
+    cents: number;
+}
